Split vision geometry tests
